program/requirements.txt
torch
torchvision
timm
pandas
numpy
matplotlib
scikit-learn
Pillow